# client/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecurringExpensePaidRequest": {
        "properties": {
          "paid_at": {
            "format": "date-time",
            "title": "Paid At",
            "type": "string"
          }
        },
        "required": [
          "paid_at"
        ],
        "title": "RecurringExpensePaidRequest",
        "type": "object"
      },
      "RecurringExpensePaidResponse": {
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/RecurringRuleResponse"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionResponse"
          }
        },
        "required": [
          "transaction",
          "rule"
        ],
        "title": "RecurringExpensePaidResponse",
        "type": "object"
      },
      "RecurringExpenseReminderListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecurringExpenseReminderResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecurringExpenseReminderListResponse",
        "type": "object"
      },
...
      "RecurringIncomeReceivedRequest": {
        "properties": {
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          }
        },
        "required": [
          "received_at"
        ],
        "title": "RecurringIncomeReceivedRequest",
        "type": "object"
      },
      "RecurringIncomeReceivedResponse": {
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/RecurringRuleResponse"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionResponse"
          }
        },
        "required": [
          "transaction",
          "rule"
        ],
        "title": "RecurringIncomeReceivedResponse",
        "type": "object"
      },
      "RecurringIncomeReminderListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecurringIncomeReminderResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecurringIncomeReminderListResponse",
        "type": "object"
      },
...
      "RecurringRuleCreateRequest": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "amount": {
            "description": "Decimal string with up to 18 digits and 4 decimal places.",
            "examples": [
              "12.3400"
            ],
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "frequency": {
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ],
            "title": "Frequency",
            "type": "string"
          },
          "interval_count": {
            "default": 1,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Interval Count",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          },
          "transaction_type": {
            "enum": [
              "income",
              "expense"
            ],
            "title": "Transaction Type",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "category_id",
          "transaction_type",
          "amount",
          "frequency",
          "start_at"
        ],
        "title": "RecurringRuleCreateRequest",
        "type": "object"
      },
      "RecurringRuleListResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecurringRuleResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "RecurringRuleListResponse",
        "type": "object"
      },
      "RecurringRuleResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "end_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "frequency": {
            "title": "Frequency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interval_count": {
            "title": "Interval Count",
            "type": "integer"
          },
          "last_paid_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Paid Period"
          },
          "last_received_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Received Period"
          },
          "last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "last_run_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Key"
          },
          "next_run_at": {
            "format": "date-time",
            "title": "Next Run At",
            "type": "string"
          },
          "paused_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused At"
          },
          "run_count": {
            "title": "Run Count",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "paused",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "transaction_type": {
            "title": "Transaction Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "category_id",
          "transaction_type",
          "amount",
          "currency",
          "description",
          "frequency",
          "interval_count",
          "timezone",
          "start_at",
          "end_at",
          "next_run_at",
          "last_run_at",
          "last_run_key",
          "last_paid_period",
          "last_received_period",
          "run_count",
          "status",
          "paused_at",
          "archived_at",
          "created_at",
          "updated_at"
        ],
        "title": "RecurringRuleResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Personal finance management API.",
    "title": "PFM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/recurring-rules": {
      "get": {
        "operationId": "list_recurring_rules_api_v1_recurring_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "active",
                "paused",
                "archived"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Recurring Rules",
        "tags": [
          "recurring"
        ]
      },
      "post": {
        "operationId": "create_recurring_rule_api_v1_recurring_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Recurring Rule",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/due-expenses": {
      "get": {
        "operationId": "list_due_recurring_expenses_api_v1_recurring_rules_due_expenses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringExpenseReminderListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Due Recurring Expenses",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/due-incomes": {
      "get": {
        "operationId": "list_due_recurring_incomes_api_v1_recurring_rules_due_incomes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringIncomeReminderListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Due Recurring Incomes",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/{rule_id}": {
      "delete": {
        "operationId": "archive_recurring_rule_api_v1_recurring_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Recurring Rule",
        "tags": [
          "recurring"
        ]
      },
      "get": {
        "operationId": "get_recurring_rule_api_v1_recurring_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Recurring Rule",
        "tags": [
          "recurring"
        ]
      },
      "patch": {
        "operationId": "update_recurring_rule_api_v1_recurring_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Recurring Rule",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/{rule_id}/paid": {
      "post": {
        "operationId": "mark_recurring_expense_paid_api_v1_recurring_rules__rule_id__paid_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringExpensePaidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringExpensePaidResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Recurring Expense Paid",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/{rule_id}/pause": {
      "post": {
        "operationId": "pause_recurring_rule_api_v1_recurring_rules__rule_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pause Recurring Rule",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/{rule_id}/received": {
      "post": {
        "operationId": "mark_recurring_income_received_api_v1_recurring_rules__rule_id__received_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringIncomeReceivedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringIncomeReceivedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Recurring Income Received",
        "tags": [
          "recurring"
        ]
      }
    },
    "/api/v1/recurring-rules/{rule_id}/resume": {
      "post": {
        "operationId": "resume_recurring_rule_api_v1_recurring_rules__rule_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume Recurring Rule",
        "tags": [
          "recurring"
        ]
      }
    },
...
  }
}